+++ tests/ui/apps/TabLayout/HorizontalTabsWithoutPreserveFocus/TabLayoutHorizontalTabsWithoutPreserveFocus-View.js
@@ -16,7 +16,6 @@ const app = (props) => (
 		<TabLayout
 			id="tabLayout"
 			orientation="horizontal"
-			preserveFocus={false}
 			style={{width: '90%'}}
 			tabSize={900}
 		>